requirements.txt
crewai==0.11.0
telethon>=1.34.0
python-dotenv>=1.0.0
langchain>=0.1.0
openai>=1.12.0
typing-extensions>=4.5.0 